fix: 템플릿 진입점을 대시보드 탭 구조에 맞춤
아바타 드롭다운과 독립 /templates 라우트는 현재 대시보드의 문서/발행 탭 기반 정보 구조와 어긋나므로 제거하고, 템플릿 목록을 기존 대시보드 탭의 세 번째 항목으로 통합했다.

Constraint: 기존 UI는 대시보드 탭 전환을 중심으로 문서와 발행을 탐색함
Rejected: 아바타 드롭다운 진입점 유지 | 핵심 문서 생명주기 기능을 계정 메뉴에 숨겨 제품 정보 구조와 맞지 않음
Confidence: high
Scope-risk: narrow
Tested: pnpm test; next build compiled successfully before existing RESEND_API_KEY page-data failure; pnpm exec tsc --noEmit shows only pre-existing SignPage/Paddle errors; git diff --cached --check
Not-tested: Browser preview interaction with real Pro account/templates data
