+++ components/user-avatar.tsx
@@ -12,7 +12,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { useLanguage } from "@/contexts/language-context";
 import { User } from "@supabase/supabase-js";
-import { LogOut, User as UserIcon, LayoutDashboard, Receipt, Settings, FileStack } from "lucide-react";
+import { LogOut, User as UserIcon, LayoutDashboard, Receipt, Settings } from "lucide-react";
 import { useState } from "react";
 import Link from "next/link";
 import { signOut } from "@/app/actions/auth-actions";
@@ -74,12 +74,6 @@ export default function UserAvatar({ user }: UserAvatarProps) {
           <Link href="/dashboard" className="cursor-pointer">
             <LayoutDashboard className="mr-2 h-4 w-4" />
             {t("dashboard.title")}
-          </Link>
-        </DropdownMenuItem>
-        <DropdownMenuItem asChild>
-          <Link href="/templates" className="cursor-pointer">
-            <FileStack className="mr-2 h-4 w-4" />
-            {t("templates.nav")}
           </Link>
         </DropdownMenuItem>
         <DropdownMenuItem asChild>
