+++ contexts/language-context.tsx
@@ -396,6 +396,7 @@ const translations: Record<Language, Record<string, string>> = {
     "dashboard.filter.completed": "완료",
     "dashboard.tabs.documents": "문서",
     "dashboard.tabs.publications": "발행",
+    "dashboard.tabs.templates": "템플릿",
     "dashboard.publications.empty.title": "발행된 문서가 없습니다",
     "dashboard.publications.empty.description": "초안 문서를 발행하여 다른 사람과 공유해보세요.",
     "dashboard.publications.empty.action": "문서 발행하기",
@@ -1565,6 +1566,7 @@ const translations: Record<Language, Record<string, string>> = {
     "dashboard.filter.completed": "Completed",
     "dashboard.tabs.documents": "Documents",
     "dashboard.tabs.publications": "Publications",
+    "dashboard.tabs.templates": "Templates",
     "dashboard.publications.empty.title": "No Published Documents",
     "dashboard.publications.empty.description": "Publish draft documents to share with others.",
     "dashboard.publications.empty.action": "Publish Documents",
